working with wishlist

+++ database/migrations/2019_11_28_071720_create_wishlists_table.php
@@ -16,7 +16,7 @@ class CreateWishlistsTable extends Migration
         Schema::create('wishlists', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->integer('user_id');
-            $table->integer('product_id');
+            $table->unsignedBigInteger('product_id');
             $table->foreign('product_id')
                 ->references('id')->on('products')
                 ->onDelete('cascade');
